feat: load program-defined metafunctions (based on #797)
A metafunction is normal Cpp2 code compiled as part of a library.
When parsing a declaration that `@`-uses the metafunction,
the library is loaded and the metafunction invoked on the declaration.

The reflection API is available by default to Cpp2 code (via `cpp2util.h`).
The implementation of the API is provided by the `cppfront` executable.
For this to work, compiling `cppfront` should export its symbols
(for an explanation, see <https://cmake.org/cmake/help/latest/prop_tgt/ENABLE_EXPORTS.html>).

The default build of `cppfront` doesn't support loading program-defined metafunctions.
In order to support loading program-defined metafunctions,
`cppfront` should be built with `CPPFRONT_LOAD_METAFUNCTION_IMPL_HEADER`
defined to a header which implements the loading functionality.
This commit includes an implementation using [Boost.DLL][].
[Boost.DLL]: https://www.boost.org/doc/libs/release/doc/html/boost_dll.html

Because cppfront doesn't perform name lookup,
metafunction names should be `@`-used unqualified
and follow C "namespacing" conventions (e.g., `@mylib_mymetafunction`).

Here is an example of program-defined metafunctions using the Boost.DLL implementation.
The commands were cleaned up from the CMake buildsystem in #797.

`metafunctions.cpp2`:
```Cpp2
greeter: (inout t: cpp2::meta::type_declaration) = {
  t.add_member($R"(say_hi: () = std::cout << "Hello, world!\nFrom (t.name())$\n";)");
}
```

`main.cpp2`:
```Cpp2
my_class: @greeter type = { }
main: ()                = my_class().say_hi();
```

Build `cppfront`:
```bash
g++ -DBOOST_ATOMIC_DYN_LINK -DBOOST_ATOMIC_NO_LIB -DBOOST_FILESYSTEM_DYN_LINK -DBOOST_FILESYSTEM_NO_LIB -DBOOST_SYSTEM_DYN_LINK -DBOOST_SYSTEM_NO_LIB -DCPPFRONT_LOAD_METAFUNCTION_IMPL_HEADER=\"reflect_load_metafunction_boost_dll.h\" -std=c++20 -o cppfront.cpp.o -c cppfront.cpp
g++ -Wl,--export-dynamic -rdynamic cppfront.cpp.o -o cppfront /usr/lib/libboost_system.so.1.83.0  /usr/lib/libboost_filesystem.so.1.83.0  /usr/lib/libboost_atomic.so.1.83.0
```

Build `metafunctions`:
```bash
./cppfront metafunctions.cpp2
g++ -std=c++20 -fPIC -o metafunctions.cpp.o -c metafunctions.cpp
g++ -fPIC -shared -Wl,-soname,libmetafunctions.so -o libmetafunctions.so metafunctions.cpp.o
```

Build and run `main`:
```bash
CPPFRONT_METAFUNCTION_LIBRARIES=libmetafunctions.so ./cppfront main.cpp2
g++ -std=c++20 -o main.cpp.o -c main.cpp
g++ main.cpp.o -o main
./main
```

Output:
```output
metafunctions.cpp2... ok (all Cpp2, passes safety checks)

main.cpp2... ok (all Cpp2, passes safety checks)

Hello, world!
From my_class
```
