+++ source/reflect_load_metafunction_never.h
@@ -1,0 +1,13 @@
+#include <functional>
+#include <string>
+#include "cpp2util.h"
+
+namespace cpp2::meta {
+
+auto load_metafunction(std::string const&)
+    -> std::function<void(type_declaration&)>
+{
+    return {};
+}
+
+}
